Move /kies utils to separate file
