+++ apps/melding-form/src/app/(map)/locatie/kies/_utils/convertWktPointToCoordinates.test.ts
@@ -1,4 +1,4 @@
-import { convertWktPointToCoordinates } from './utils'
+import { convertWktPointToCoordinates } from './convertWktPointToCoordinates'
 
 describe('convertWktPointToCoordinates', () => {
   it('should convert a WKT point to lat lng location with comma or space as separation', () => {
