+++ apps/melding-form/src/app/(map)/locatie/kies/_utils/convertWktPointToCoordinates.ts
